convert interim faq and deriv and dmt5 to ts
