+++ src/pages/interim/deriv/_hero.tsx
