+++ src/pages/interim/deriv/index.tsx
@@ -1,5 +1,4 @@
 import React from 'react'
-import PropTypes from 'prop-types'
 import Loadable from '@loadable/component'
 import Everything from '../_everything'
 import Hero from './_hero'
@@ -11,18 +10,14 @@ const LoveTrading = Loadable(() => import('./_love-trading'))
 
 const Interim = () => {
     return (
-        <Layout type="interim" interim_type="dmt5">
-            <SEO title={localize('Interim | DMT5')} no_index />
+        <Layout type="interim" interim_type="deriv">
+            <SEO title={localize('Interim | Deriv')} no_index />
             <Hero />
             <Everything />
             <FAQ />
             <LoveTrading />
         </Layout>
     )
-}
-
-Interim.propTypes = {
-    referrer: PropTypes.string,
 }
 
 export default WithIntl()(Interim)
